--- a/app/controllers/ontologies_controller.rb
+++ b/app/controllers/ontologies_controller.rb
@@ -32,45 +32,30 @@ class OntologiesController < ApplicationController
   def index
     @categories = LinkedData::Client::Models::Category.all(display_links: false, display_context: false)
     @groups = LinkedData::Client::Models::Group.all(display_links: false, display_context: false)
+    
     @filters = ontology_filters_init(@categories, @groups)
     init_filters(params)
     render 'ontologies/browser/browse'
   end
 
   def ontologies_filter
-    @ontologies = submissions_paginate_filter(params)
-    @object_count = count_objects(@ontologies)
-
-    update_filters_counts = @object_count.map do |section, values_count|
-      values_count.map do |value, count|
-        replace("count_#{section}_#{value}") do
-          helpers.turbo_frame_tag("count_#{section}_#{value}") do
-            helpers.content_tag(:span, count.to_s, class: "hide-if-loading #{count.zero? ? 'disabled' : ''}")
+    @ontologies, @count, @count_objects, @request_params = submissions_paginate_filter(params)
+    if @page.page.eql?(1)
+      streams = [prepend("ontologies_list_view-page-#{@page.page}", partial: 'ontologies/browser/ontologies')]
+      streams += @count_objects.map do |section, values_count|
+        values_count.map do |value, count|
+          replace("count_#{section}_#{value}") do
+            helpers.turbo_frame_tag("count_#{section}_#{value}") do
+              helpers.content_tag(:span, count.to_s, class: "hide-if-loading #{count.zero? ? 'disabled' : ''}")
+            end
           end
         end
-      end
-    end.flatten
+      end.flatten
+    else
+      streams = [replace("ontologies_list_view-page-#{@page.page}", partial: 'ontologies/browser/ontologies')]
+    end
 
-    count_streams = [
-      replace('ontologies_filter_count_request') do
-        helpers.content_tag(:p, class: "browse-desc-text", style: "margin-bottom: 12px !important;") { t("ontologies.showing_ontologies_size", ontologies_size: @ontologies.size, analytics_size: @analytics.keys.size) }
-      end
-    ] + update_filters_counts
-
-    streams = if params[:page].nil?
-                [
-                  prepend('ontologies_list_container', partial: 'ontologies/browser/ontologies'),
-                  prepend('ontologies_list_container') {
-                    helpers.turbo_frame_tag("ontologies_filter_count_request") do
-                      helpers.browser_counter_loader
-                    end
-                  }
-                ]
-              else
-                [replace("ontologies_list_view-page-1", partial: 'ontologies/browser/ontologies')]
-              end
-
-    render turbo_stream: streams + count_streams
+    render turbo_stream: streams
   end
 
   def classes
@@ -553,13 +538,6 @@ class OntologiesController < ApplicationController
     properties.map { |x| [x.to_s, [sub.send(x.to_s), custom_labels[x.to_sym]]] }.to_h
   end
 
-  def get_metrics_hash
-    metrics_hash = {}
-    # TODO: Metrics do not return for views on the backend, need to enable include_views param there
-    @metrics = LinkedData::Client::Models::Metrics.all(include_views: true)
-    @metrics.each { |m| metrics_hash[m.links['ontology']] = m }
-    return metrics_hash
-  end
 
   def determine_layout
     case action_name