--- a/app/helpers/fair_score_helper.rb
+++ b/app/helpers/fair_score_helper.rb
@@ -11,16 +11,24 @@ module FairScoreHelper
   def get_fairness_service_url(apikey = user_apikey)
     "#{$FAIRNESS_URL}?portal=#{$HOSTNAME.split('.')[0]}#{apikey.nil? || apikey.empty? ? '' : "&apikey=#{apikey}"}"
   end
+
   def get_fairness_json(ontologies_acronyms, apikey = user_apikey)
-    begin
-      conn = Faraday.new do |conn|
-        conn.options.timeout = 30
+    Rails.cache.fetch("fairness-#{ontologies_acronyms.gsub(',', '-')}", expires: 24.hours) do
+      begin
+        out = {}
+        time = Benchmark.realtime do
+          conn = Faraday.new do |conn|
+            conn.options.timeout = 30
+          end
+          response = conn.get(get_fairness_service_url(apikey) + "&ontologies=#{ontologies_acronyms}&combined")
+          out = MultiJson.load(response.body.force_encoding('ISO-8859-1').encode('UTF-8'))
+        end
+        puts "Call fairness service for: #{ontologies_acronyms} (#{time}s)"
+      rescue
+        Rails.logger.warn t('fair_score.fairness_unreachable_warning')
       end
-      response = conn.get(get_fairness_service_url(apikey) + "&ontologies=#{ontologies_acronyms}&combined")
-      MultiJson.load(response.body.force_encoding('ISO-8859-1').encode('UTF-8'))
-    rescue
-      Rails.logger.warn t('fair_score.fairness_unreachable_warning')
-      {}
+
+      out
     end
   end
 
@@ -72,7 +80,7 @@ module FairScoreHelper
   end
 
   def get_not_obtained_score(fair_scores_data, index)
-      fair_scores_data[:criteria][:portalMaxCredits][index] - fair_scores_data[:criteria][:scores][index]
+    fair_scores_data[:criteria][:portalMaxCredits][index] - fair_scores_data[:criteria][:scores][index]
   end
 
   def get_not_obtained_score_normalized(fair_scores_data, index)
@@ -84,7 +92,7 @@ module FairScoreHelper
     elsif score_rest.zero?
         100 - fair_scores_data[:criteria][:normalizedScores][index]
     else
-         0
+      0
     end
 
   end