--- a/app/views/ontologies/browser/_ontologies.html.haml
+++ b/app/views/ontologies/browser/_ontologies.html.haml
@@ -1,7 +1,11 @@
 = render InfiniteScrollComponent.new(id: 'ontologies_list',
                                     collection: @ontologies,
-                                    next_url: ontologies_filter_url(@filters, page: @page.nextPage),
+                                    next_url: ontologies_filter_url(@request_params, page: @page.nextPage),
                                     current_page: @page.page, next_page: @page.nextPage) do |c|
+
+  - if @page.page.eql?(1)
+    = content_tag(:p, class: "browse-desc-text", style: "margin-bottom: 12px !important;") { "Showing #{@count} of #{@total_ontologies} (#{sprintf("%.2f", @time)}s)" }
+
   - ontologies = c.collection
   - ontologies.each do |ontology|
     = render OntologyBrowseCardComponent.new(ontology: ontology)