… app/views/ontologies/browser/browse.html.haml
… app/views/ontologies/ontologies_selector/ontologies_selector_results.html.haml
… app/views/submissions/_form_content.html.haml