### Context 
See https://github.com/agroportal/project-management/issues/241 and https://github.com/ontoportal-lirmm/bioportal_web_ui/issues/425

### Changes
* Optimize fetched attributes in the browse page in API mode (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/e554b9dcb490a5426f32a8aebeac43b71cfa5a84)
* Paginate the browse page 
* Update ontologies browse to use API or Index (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/9c659e5414fd357e61c70faa008673728d74c28d)
* Add cache for the fair_score and submission endpoint (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/17888ef535ca9c80f1d37c256ea1f456a50a7766)
* Internationalize a sentence in the browse and ontologies selector views (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/2bdb90cb32b5c39c0e4873bae7e096083d564dde)
* Fix browse hasFormality and isOfType filters (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/53cc4b4857d0fd4e5f105da10ea2448fa4c30a4a)
* Prioritize API filtring over the index in the browse page (https://github.com/ontoportal-lirmm/bioportal_web_ui/pull/572/commits/5d2c2f0a84d54a8872b571cf4abe722f27c10cc5)


### Benchmarks 

#### Different options + with or without cache
  | Current code - Triple store (without cache) | New code - Triple store (without cache) | Current code- Triple store (with cache) | New code -  SOLR | New code -  Triple store (with cache)
-- | -- | -- | -- | -- | --
Browse all (70% usage) | 3.9s | 2.98s | 2s | 0.93s | 0.21s
Search ACRONYM: AGROVOC (15% usage) | 3.6s | 3.01s | 1.8s | 0.48s | 0.12s
Search Name: `INRAE Thesaurus` (10% usage) (| 3.2s | 2.66s | 1.74s | 0.45s | 0.10s
Search Description: `This vocabulary is a first formalization of the INSPIRE` (5% usage) | 3.5s | 2.55s | 2.1s | 0.73s | 0.15s

#### Different options summary

  | Current code - triple store | New code - Triple store  | New code -  SOLR
-- | -- | -- | --
Browse all cold run | 3.9s | 2.98s | 0.93s
Browse all cached run | 2s | 0.21s | 0.93s
  |   |   |  
2 X Browse all | 3.9s  + 2s  = 5.9s | 2.98s +  0.21s  = 3,1s | 0.93s + 0.93s = 1,86s
3 X Browse all | 3.9s  + 2s + 2s = 7.9s | 2.98s +  0.21s +  0.21s = 3,4s | 0.93s + 0.93s + 0.93s = 2,7s
4 X Browse all | 3.9s  + 2s + 2s +2s = 9.9s | 2.98s +  0.21s +  0.21s + 0,21s = 3,6s | 0.93s + 0.93s + 0.93s  + 0,93s = 3,7s
  |   |   |  
Advantage | Time dependent of request complexity    | Less complex + Federation compliant | Better cold start + Time dependent of request complexity  + better result ranking
Disadvantage | More complex + Federation no compliant + 4store does not support pagination | Cold start | More complex + Federation no compliant