@jonquet 2024-04-27
Congrats on this study, analysis and implementation. This is a model of how technical choices should be taken!